requirements.txt
discord.py
python-dotenv
groq
playwright
fastmcp
duckduckgo-search
playwright
pytest-playwright
